app/requirements.txt
python-binance==1.0.15
rich==11.0.0
redis==4.1.1
Flask==2.0.2
